update/det-metrics code correction based on review
